requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pymongo>=4.6
SQLAlchemy>=2.0
apache-airflow>=2.8
python-dotenv>=1.0
matplotlib>=3.8
seaborn>=0.13